Fixed Change Requests for PR
