+++ src/genophenocorr/model/__init__.py
@@ -10,13 +10,13 @@ from ._gt import Genotype, Genotypes, Genotyped
 from ._phenotype import Phenotype, Disease
 from ._protein import FeatureInfo, FeatureType, ProteinFeature, ProteinMetadata
 from ._tx import TranscriptCoordinates
-from ._variant import VariantCoordinates, TranscriptAnnotation, TranscriptInfoAware, Variant, VariantFormatter
+from ._variant import VariantCoordinates, TranscriptAnnotation, TranscriptInfoAware, Variant
 from ._variant_effects import VariantEffect
 
 __all__ = [
     'Cohort', 'Patient', 'SampleLabels',
     'Phenotype', 'Disease',
-    'Variant', 'VariantCoordinates', 'Genotype', 'Genotypes', 'Genotyped', 'VariantFormatter',
+    'Variant', 'VariantCoordinates', 'Genotype', 'Genotypes', 'Genotyped', 
     'TranscriptAnnotation', 'VariantEffect', 'TranscriptInfoAware', 'TranscriptCoordinates',
     'ProteinMetadata', 'ProteinFeature', 'FeatureInfo', 'FeatureType',
 ]
