+++ src/genophenocorr/model/_variant.py
@@ -7,7 +7,7 @@ import hpotk
 from .genome import Region, GenomicRegion
 from ._gt import Genotyped, Genotypes
 from ._variant_effects import VariantEffect
-from ._formatter import Formatter
+from ..view._formatter import Formatter
 
 
 class TranscriptInfoAware(metaclass=abc.ABCMeta):
@@ -395,6 +395,20 @@ class FunctionalAnnotationAware(metaclass=abc.ABCMeta):
             if tx_ann.transcript_id == transcript_id:
                 return tx_ann.hgvs_cdna
         return None
+    
+    def get_preferred_tx_annotation(self) -> typing.Optional[TranscriptAnnotation]:
+        """A TranscriptAnnotation that represents the result of the functional annotation
+        of a variant with respect to the preferred transcript of a gene. Will return None if there is 
+        no preferred transcript found. 
+
+        Returns:
+            typing.Optional[TranscriptAnnotation]: A TranscriptAnnotation object representing the 
+                                                    preferred transcript if one can be found
+        """
+        for tx in self.tx_annotations:
+            if tx.is_preferred:
+                return tx
+        return None
 
 
 class Variant(VariantCoordinateAware, FunctionalAnnotationAware, Genotyped):
@@ -460,25 +474,6 @@ class Variant(VariantCoordinateAware, FunctionalAnnotationAware, Genotyped):
         return self._tx_annotations
 
     @property
-    def mane_tx_annotation(self) -> typing.Optional[TranscriptAnnotation]:
-        """A TranscriptAnnotation that represents the result of the functional annotation
-        of a variant with respect to the MANE transcript of a gene. Will return None if there is 
-        no MANE transcript found. 
-        
-        @ielis (Daniel): Would this be a function for FunctionalAnnotationAware as well? Or should
-        it be left here? 
-
-        Returns:
-            typing.Optional[TranscriptAnnotation]: A TranscriptAnnotation object representing the 
-                                                    MANE transcript if one can be found
-        """
-        for tx in self.tx_annotations:
-            if tx.is_preferred:
-                return tx
-        return None
-
-
-    @property
     def genotypes(self) -> Genotypes:
         return self._gts
 
@@ -500,28 +495,3 @@ class Variant(VariantCoordinateAware, FunctionalAnnotationAware, Genotyped):
                 f"tx_annotations:{self.tx_annotations}, "
                 f"genotypes:{self.genotypes})")
 
-class VariantFormatter(Formatter[Variant]):
-    """
-    A class that can be used to format a `Variant` to a human readable string
-    """
-    def format_as_string(self, item: Variant, tx_id: typing.Optional[str] = None) -> str:
-        """ 
-        Args:
-            item (Variant): An object of class `Variant` representing a variant.
-            tx_id (typing.Optional[str]): The transcript ID, needed if using a transcript other than the MANE transcript.
-
-        Returns:
-            str: A human readable string for the variant.
-        """
-        if tx_id is not None:
-            transcript = item.get_tx_anno_by_tx_id(tx_id)
-        else:
-            transcript = item.mane_tx_annotation
-        if transcript is not None:
-            if len(transcript.hgvs_cdna) > 50:
-                return "Long HGVS"
-            return transcript.hgvs_cdna
-        elif item.variant_coordinates.variant_key is not None:
-            return item.variant_coordinates.variant_key
-        else:
-            return f"Variant {item} has no string format."
