+++ src/genophenocorr/view/__init__.py
@@ -5,11 +5,13 @@ from ._protein_visualizable import ProteinVisualizable
 from ._stats import StatsViewer
 from ._txp import VariantTranscriptVisualizer
 from ._protein_visualizer import ProteinVisualizer
+from ._formatter import VariantFormatter
 
 __all__ = [
     'CohortViewable',
     'ProteinVisualizer', 'ProteinVisualizable', 'ProteinViewable',
     'DiseaseViewable',
     'StatsViewer',
-    'VariantTranscriptVisualizer'
+    'VariantTranscriptVisualizer',
+    'VariantFormatter'
 ]
